requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.0
SQLAlchemy==2.0.25
psycopg2-binary==2.9.5
gunicorn==21.2.0
uvicorn==0.23.2
starlette==0.37.2
jinja2==3.1.3
aiofiles==23.2.1
Pillow 